+++ scripts/bench.py
@@ -192,7 +192,7 @@ def run_ecma_tonality() -> np.ndarray:
 
 
 def run_vibration_weighting() -> np.ndarray:
-    from phonometry.human_vibration import apply_weighting
+    from phonometry.vibration.human_vibration import apply_weighting
 
     sig, fs = vibration_signal()
     out = np.asarray(apply_weighting(sig, fs, "Wd"), dtype=np.float64)
