+++ scripts/generate_graphs.py
@@ -3446,7 +3446,7 @@ def generate_wind_turbine_tonality(output_dir: str) -> None:
     """IEC 61400-11 wind-turbine tonal audibility: narrowband spectrum + masking."""
     print("Generating wind_turbine_tonality...")
     from phonometry import wind_turbine_tonality
-    from phonometry.wind_turbine_noise import _critical_band_edges
+    from phonometry.environmental.wind_turbine_noise import _critical_band_edges
 
     # A narrowband spectrum: a shaped broadband floor with a blade-passing-style
     # tone near 200 Hz, at 2 Hz resolution.
@@ -5746,7 +5746,7 @@ def generate_impulse_prominence(output_dir: str) -> None:
         impulse_prominence,
         predicted_prominence,
     )
-    from phonometry.impulse_prominence import ADJUSTMENT_THRESHOLD
+    from phonometry.environmental.impulse_prominence import ADJUSTMENT_THRESHOLD
 
     fig, (ax_p, ax_k) = plt.subplots(1, 2, figsize=(12.5, 5.4))
 
@@ -5850,7 +5850,7 @@ def generate_multiple_shock(output_dir: str) -> None:
         injury_risk,
         seat_to_spine_transfer,
     )
-    from phonometry.multiple_shock_vibration import (
+    from phonometry.vibration.multiple_shock_vibration import (
         MZ_MALE,
         RISK_THRESHOLDS_MALE,
     )
