+++ src/phonometry/_plot/common.py
@@ -31,28 +31,16 @@ if TYPE_CHECKING:
     from matplotlib.axes import Axes
     from matplotlib.container import BarContainer
 
-    from ..mechanical_mobility import MobilityResult
-    from ..transfer_stiffness import TransferStiffnessResult
     from ..building.insulation import (
         ImpactRatingResult,
         WeightedRatingResult,
     )
-    from ..outdoor_propagation import OutdoorAttenuation
     from ..room.room_acoustics import RoomAcousticsResult
-    from ..human_vibration import (
-        DailyVibrationExposure,
-        WeightedSpectrum,
-        WeightingResponse,
-    )
-    from ..multiple_shock_vibration import MultipleShockResult
-    from ..environmental_measurement import TonalAssessmentResult
-    from ..impulse_prominence import ImpulseProminenceResult
     from ..distortion import HarmonicDistortionResult
     from ..frequency_response import FrequencyResponseResult
     from ..ship_radiated_noise import ShipSourceLevelResult
     from ..pile_driving_noise import PileStrikeResult
     from ..aircraft_noise import EPNLResult
-    from ..wind_turbine_noise import WindTurbineTonalityResult
     from ..underwater_sound_speed import SoundSpeedProfile
     from ..underwater_propagation import TransmissionLossResult
     from ..sonar_equation import SonarEquationResult
@@ -477,35 +465,6 @@ def plot_epnl(result: "EPNLResult", ax: Axes | None = None, **kwargs: Any) -> Ax
     return ax
 
 
-def plot_wind_turbine_tonality(
-    result: "WindTurbineTonalityResult", ax: Axes | None = None, **kwargs: Any
-) -> Axes:
-    """Narrowband spectrum with the critical band, masking level and the tone.
-
-    :param result: A :class:`~phonometry.wind_turbine_noise.WindTurbineTonalityResult`.
-    :param ax: Existing axes, or ``None`` to create a figure.
-    :param kwargs: Forwarded to the spectrum ``plot`` call.
-    :return: The axes.
-    """
-    from ..wind_turbine_noise import _critical_band_edges
-
-    ax = ax if ax is not None else _new_axes()
-    freqs = np.asarray(result.frequencies, dtype=np.float64)
-    levels = np.asarray(result.levels, dtype=np.float64)
-    fc = result.tone_frequency
-    lo, hi = _critical_band_edges(fc)
-    ax.plot(freqs, levels, **{"color": _C_PRIMARY, "lw": 1.0, "label": "Narrowband spectrum", **kwargs})
-    ax.axvspan(lo, hi, color=_C_TERTIARY, alpha=0.12, label="Critical band")
-    ax.axhline(result.masking_level, color=_C_MUTED, ls="--", lw=1.0,
-               label=f"Masking level ({result.masking_level:.1f} dB)")
-    ax.plot([fc], [result.tone_level], "o", color=_C_REFERENCE,
-            label=f"Tone ({result.tone_level:.1f} dB)")
-    ax.set_xlabel("Frequency [Hz]")
-    ax.set_ylabel("Level [dB]")
-    ax.set_title(f"IEC 61400-11 tonal audibility ΔLₐ = {result.tonal_audibility:.1f} dB")
-    ax.grid(True, alpha=0.3)
-    ax.legend(loc=_LEGEND_UPPER_RIGHT, fontsize="small")
-    return ax
 
 
 def plot_sound_speed_profile(
@@ -1266,119 +1225,10 @@ def plot_excitation(
 # ---------------------------------------------------------------------------
 
 
-def plot_vibration_weighting(
-    result: "WeightingResponse", ax: Axes | None = None, **kwargs: Any
-) -> Axes:
-    """Frequency-weighting factor (dB) versus frequency (ISO 8041-1).
-
-    :param result: A
-        :class:`~phonometry.human_vibration.WeightingResponse` exposing
-        ``name``, ``frequencies`` and ``magnitude_db``.
-    :param ax: Existing axes, or ``None`` to create a figure.
-    :param kwargs: Forwarded to the weighting curve ``plot`` call.
-    :return: The axes.
-    """
-    ax = ax if ax is not None else _new_axes()
-    freqs = np.asarray(result.frequencies, dtype=np.float64)
-    mag_db = np.asarray(result.magnitude_db, dtype=np.float64)
-    kwargs.setdefault("color", _C_PRIMARY)
-    ax.semilogx(freqs, mag_db, **kwargs)
-    ax.set_xlabel("Frequency [Hz]")
-    ax.set_ylabel("Weighting factor [dB]")
-    ax.set_title(f"Frequency weighting {result.name} (ISO 8041-1)")
-    ax.grid(True, which="both", alpha=0.3)
-    return ax
 
 
-def plot_weighted_spectrum(
-    result: "WeightedSpectrum", ax: Axes | None = None, **kwargs: Any
-) -> Axes:
-    """Unweighted vs weighted one-third-octave acceleration spectrum.
-
-    Draws the measured band accelerations and, overlaid, the weighted band
-    contributions ``W_i*a_i``; the overall ``a_w`` is annotated in the title.
-
-    :param result: A
-        :class:`~phonometry.human_vibration.WeightedSpectrum` exposing
-        ``frequencies``, ``band_accelerations``, ``weighted``, ``overall`` and
-        ``weighting_name``.
-    :param ax: Existing axes, or ``None`` to create a figure.
-    :param kwargs: Forwarded to the weighted (primary) bars.
-    :return: The axes.
-    """
-    ax = ax if ax is not None else _new_axes()
-    freqs = np.asarray(result.frequencies, dtype=np.float64)
-    raw = np.asarray(result.band_accelerations, dtype=np.float64)
-    weighted = np.asarray(result.weighted, dtype=np.float64)
-    positions = _band_axis(ax, freqs)
-    width = 0.4
-    # The weighted bars are the primary artist; forward user kwargs there.
-    kwargs.setdefault("color", _C_PRIMARY)
-    ax.bar(
-        positions - width / 2, raw, width, color=_C_MUTED, label="Unweighted $a_i$"
-    )
-    ax.bar(
-        positions + width / 2,
-        weighted,
-        width,
-        label=f"Weighted $W_i a_i$ ({result.weighting_name})",
-        **kwargs,
-    )
-    ax.set_ylabel(r"r.m.s. acceleration [m/s$^2$]")
-    # Wh is the hand-arm weighting of ISO 5349-1; the others (Wk, Wd, Wm...)
-    # are the whole-body weightings of ISO 2631.
-    designation = "ISO 5349-1" if str(result.weighting_name) == "Wh" else "ISO 2631"
-    ax.set_title(
-        f"{designation} weighted acceleration spectrum  "
-        f"($a_w$ = {float(result.overall):.3f} " r"m/s$^2$)"
-    )
-    ax.legend(loc="best", fontsize="small")
-    ax.grid(True, axis="y", alpha=0.3)
-    return ax
 
 
-def plot_daily_exposure(
-    result: "DailyVibrationExposure", ax: Axes | None = None, **kwargs: Any
-) -> Axes:
-    """Partial daily exposures against the EAV / ELV (Directive 2002/44/EC).
-
-    Draws one bar per operation (its partial exposure ``A_i(8)``), a combined
-    ``A(8)`` bar, and the exposure action and limit value as horizontal lines.
-
-    :param result: A
-        :class:`~phonometry.human_vibration.DailyVibrationExposure` exposing
-        ``labels``, ``partials``, ``a8`` and ``assessment``.
-    :param ax: Existing axes, or ``None`` to create a figure.
-    :param kwargs: Forwarded to the exposure :meth:`~matplotlib.axes.Axes.bar`.
-    :return: The axes.
-    """
-    ax = ax if ax is not None else _new_axes()
-    partials = np.asarray(result.partials, dtype=np.float64)
-    labels = [*result.labels, "A(8)"]
-    values = [*partials.tolist(), float(result.a8)]
-    positions = np.arange(len(values), dtype=np.float64)
-    colors = [_C_MUTED] * partials.size + [_C_PRIMARY]
-    kwargs.setdefault("color", colors)
-    ax.bar(positions, values, **kwargs)
-    ax.set_xticks(positions)
-    ax.set_xticklabels(labels, rotation=45, ha="right")
-    ax.set_ylabel(r"Vibration exposure A(8) [m/s$^2$]")
-
-    assessment = result.assessment
-    eav = float(assessment.action_value)
-    elv = float(assessment.limit_value)
-    ax.axhline(eav, color=_C_SECONDARY, ls="--", label=f"EAV = {eav:g}")
-    ax.axhline(elv, color=_C_REFERENCE, ls="--", label=f"ELV = {elv:g}")
-    top = max(elv, float(np.max(values))) * 1.15
-    ax.set_ylim(0.0, top)
-    kind = str(assessment.kind).upper()
-    ax.set_title(
-        f"Directive 2002/44/EC daily {kind} exposure  "
-        f"(A(8) = {float(result.a8):.2f} " rf"m/s$^2$, {assessment.zone})"
-    )
-    ax.legend(loc="best", fontsize="small")
-    ax.grid(True, axis="y", alpha=0.3)
-    return ax
 
 
 # ---------------------------------------------------------------------------
@@ -1411,79 +1261,6 @@ def plot_daily_exposure(
 # ---------------------------------------------------------------------------
 
 
-def plot_impulse_prominence(
-    result: "ImpulseProminenceResult", ax: Axes | None = None, **kwargs: Any
-) -> Axes:
-    """Adjustment curve ``KI(P)`` with the candidate impulses marked.
-
-    :param result: An :class:`~phonometry.impulse_prominence.ImpulseProminenceResult`.
-    :param ax: Existing axes, or ``None`` to create a figure.
-    :param kwargs: Forwarded to the impulses ``scatter``.
-    :return: The axes.
-    """
-    from ..impulse_prominence import ADJUSTMENT_THRESHOLD, impulse_adjustment
-
-    ax = ax if ax is not None else _new_axes()
-    per = np.asarray(result.per_impulse, dtype=np.float64)
-    per_max = float(per.max()) if per.size else 0.0
-    p_max = max(per_max, result.prominence, 15.0) + 1.0
-    grid = np.linspace(0.0, p_max, 200)
-    ax.plot(grid, impulse_adjustment(grid), color=_C_PRIMARY,
-            label=r"$K_I = 1.8\,(P-5)$")
-    ax.axvline(ADJUSTMENT_THRESHOLD, color=_C_MUTED, ls=":",
-               label=f"threshold $P = {ADJUSTMENT_THRESHOLD:g}$")
-
-    kwargs.setdefault("color", _C_PRIMARY_LIGHT)
-    kwargs.setdefault("zorder", 3)
-    ax.scatter(per, impulse_adjustment(per), label="Impulses", **kwargs)
-    ax.scatter([result.prominence], [result.adjustment], color=_C_REFERENCE,
-               zorder=4, s=90, marker="*",
-               label=f"Governing  P = {result.prominence:.2f},  "
-                     f"$K_I$ = {result.adjustment:.1f} dB")
-    ax.set_xlabel("Predicted prominence $P$")
-    ax.set_ylabel("Adjustment $K_I$ [dB]")
-    ax.set_title("NT ACOU 112 — impulse adjustment to $L_{Aeq}$")
-    ax.set_ylim(bottom=0.0)
-    ax.legend(loc="upper left", fontsize="small")
-    ax.grid(True, alpha=0.3)
-    return ax
-
-
-def plot_tonal_adjustment(
-    result: "TonalAssessmentResult", ax: Axes | None = None, **kwargs: Any
-) -> Axes:
-    """Tonal adjustment curve ``Kt(ΔLta)`` with the assessed tone marked.
-
-    :param result: A
-        :class:`~phonometry.environmental_measurement.TonalAssessmentResult`.
-    :param ax: Existing axes, or ``None`` to create a figure.
-    :param kwargs: Forwarded to the assessed-tone ``scatter``.
-    :return: The axes.
-    """
-    from ..environmental_measurement import tonal_adjustment
-
-    ax = ax if ax is not None else _new_axes()
-    top = max(result.audibility, 12.0) + 1.0
-    grid = np.linspace(0.0, top, 200)
-    curve = np.array([tonal_adjustment(d) for d in grid], dtype=np.float64)
-    ax.plot(grid, curve, color=_C_PRIMARY, label=r"$K_t(\Delta L_{ta})$")
-    ax.axvline(4.0, color=_C_MUTED, ls=":", label=r"knees $\Delta L_{ta}=4,\,10$ dB")
-    ax.axvline(10.0, color=_C_MUTED, ls=":")
-
-    kwargs.setdefault("color", _C_REFERENCE)
-    kwargs.setdefault("zorder", 4)
-    kwargs.setdefault("s", 90)
-    kwargs.setdefault("marker", "*")
-    ax.scatter([result.audibility], [result.adjustment],
-               label=rf"$\Delta L_{{ta}}$ = {result.audibility:.1f} dB,  "
-                     rf"$K_t$ = {result.adjustment:.1f} dB", **kwargs)
-    ax.set_xlabel(r"Tonal audibility $\Delta L_{ta}$ [dB]")
-    ax.set_ylabel("Tonal adjustment $K_t$ [dB]")
-    ax.set_title("ISO 1996-2 tonal adjustment")
-    ax.set_ylim(bottom=0.0)
-    ax.legend(loc="upper left", fontsize="small")
-    ax.grid(True, alpha=0.3)
-    return ax
 
 
 
@@ -1494,55 +1271,10 @@ def plot_tonal_adjustment(
 
 
 
-def plot_mobility(
-    result: "MobilityResult", ax: Axes | None = None, **kwargs: Any
-) -> Axes:
-    """Mobility magnitude ``|Y(f)|`` on log-log axes (ISO 7626-1).
-
-    :param result: A :class:`~phonometry.mechanical_mobility.MobilityResult`.
-    :param ax: Existing axes, or ``None`` to create a figure.
-    :param kwargs: Forwarded to the magnitude ``plot``.
-    :return: The axes.
-    """
-    ax = ax if ax is not None else _new_axes()
-    freq = np.asarray(result.frequencies, dtype=np.float64)
-    mag = np.asarray(result.magnitude, dtype=np.float64)
-    kwargs.setdefault("color", _C_PRIMARY)
-    label = "driving-point mobility" if result.driving_point else "transfer mobility"
-    ax.loglog(freq, mag, label=label, **kwargs)
-    # Mark the mobility peak (a resonance for a driving-point FRF).
-    peak = int(np.argmax(mag))
-    ax.plot(freq[peak], mag[peak], "o", color=_C_REFERENCE, zorder=5,
-            label=f"peak at {freq[peak]:.1f} Hz")
-    ax.set_xlabel("Frequency [Hz]")
-    ax.set_ylabel("Mobility $|Y|$ [m/(N·s)]")
-    ax.set_title("ISO 7626-1 mechanical mobility")
-    ax.legend(loc="best", fontsize="small")
-    ax.grid(True, which="both", alpha=0.3)
-    return ax
 
 
-def plot_transfer_stiffness(
-    result: "TransferStiffnessResult", ax: Axes | None = None, **kwargs: Any
-) -> Axes:
-    """Dynamic transfer stiffness level ``L_k(f)`` on a log-frequency axis.
 
-    :param result: A :class:`~phonometry.transfer_stiffness.TransferStiffnessResult`.
-    :param ax: Existing axes, or ``None`` to create a figure.
-    :param kwargs: Forwarded to the level ``plot``.
-    :return: The axes.
-    """
-    ax = ax if ax is not None else _new_axes()
-    freq = np.asarray(result.frequencies, dtype=np.float64)
-    level = np.asarray(result.level, dtype=np.float64)
-    kwargs.setdefault("color", _C_PRIMARY)
-    ax.semilogx(freq, level, label=r"$L_k = 20\,\lg(|k_{2,1}|/k_0)$", **kwargs)
-    ax.set_xlabel("Frequency [Hz]")
-    ax.set_ylabel(r"Transfer stiffness level $L_k$ [dB re 1 N/m]")
-    ax.set_title("ISO 10846 dynamic transfer stiffness")
-    ax.legend(loc="best", fontsize="small")
-    ax.grid(True, which="both", alpha=0.3)
-    return ax
+
 
 
 def _plot_band_level_bars(
@@ -1585,43 +1317,6 @@ def _plot_band_level_bars(
 
 
 
-def plot_multiple_shock(
-    result: "MultipleShockResult", ax: Axes | None = None, **kwargs: Any
-) -> Axes:
-    """Injury-probability curve ``P(R)`` with this assessment's ``R`` marked.
-
-    :param result: A :class:`~phonometry.multiple_shock_vibration.MultipleShockResult`.
-    :param ax: Existing axes, or ``None`` to create a figure.
-    :param kwargs: Forwarded to the ``R`` marker ``scatter``.
-    :return: The axes.
-    """
-    from ..multiple_shock_vibration import injury_probability
-
-    ax = ax if ax is not None else _new_axes()
-    r10, r50, r90 = result.risk_thresholds
-    r_max = max(result.risk, r90) * 1.3
-    grid = np.linspace(0.0, r_max, 240)
-    prob = np.asarray(injury_probability(grid, sex=result.sex), dtype=np.float64)
-    ax.plot(grid, 100.0 * prob, color=_C_PRIMARY,
-            label=r"$\Pi(R) = 1 - e^{-(R/\alpha)^{\beta}}$")
-    for level, r_val in zip((10, 50, 90), (r10, r50, r90)):
-        ax.axhline(level, color=_C_MUTED, ls=":", lw=0.8)
-        ax.plot([r_val, r_val], [0.0, level], color=_C_MUTED, ls=":", lw=0.8)
-
-    kwargs.setdefault("color", _C_REFERENCE)
-    kwargs.setdefault("zorder", 4)
-    kwargs.setdefault("s", 90)
-    ax.scatter([result.risk], [100.0 * result.probability],
-               label=f"$R$ = {result.risk:.2f},  $\\Pi$ = "
-                     f"{100.0 * result.probability:.0f} %", **kwargs)
-    ax.set_xlabel("Stress variable $R$")
-    ax.set_ylabel("Probability of lumbar injury [%]")
-    ax.set_title(f"ISO 2631-5 injury probability — {result.sex}")
-    ax.set_xlim(left=0.0)
-    ax.set_ylim(0.0, 100.0)
-    ax.legend(loc="lower right", fontsize="small")
-    ax.grid(True, alpha=0.3)
-    return ax
 
 
 # ---------------------------------------------------------------------------
@@ -1645,56 +1340,6 @@ def plot_multiple_shock(
 # ---------------------------------------------------------------------------
 
 
-def plot_outdoor_attenuation(
-    result: "OutdoorAttenuation", ax: Axes | None = None, **kwargs: Any
-) -> Axes:
-    """Stacked per-band attenuation terms with the total overlaid (ISO 9613-2).
-
-    The divergence, atmospheric, ground and barrier terms are stacked per
-    octave band on separate positive and negative baselines (the ground
-    effect can be a net *gain*), and the total attenuation ``A`` is drawn
-    as the primary marker line on top.
-
-    :param result: An
-        :class:`~phonometry.outdoor_propagation.OutdoorAttenuation`.
-    :param ax: Existing axes, or ``None`` to create a figure.
-    :param kwargs: Forwarded to the total-attenuation ``plot`` call.
-    :return: The axes.
-    """
-    ax = ax if ax is not None else _new_axes()
-    freqs = np.asarray(result.frequencies, dtype=np.float64)
-    positions = _band_axis(ax, freqs)
-    n = freqs.size
-
-    # Separate positive and negative cumulative baselines so a negative term
-    # stacks below zero instead of being drawn on top of the previous bars;
-    # the signed heights sum to a_total.
-    pos_bottom = np.zeros(n)
-    neg_bottom = np.zeros(n)
-    terms = (
-        (result.a_div, _C_PRIMARY, "$A_{div}$ — divergence"),
-        (result.a_atm, _C_TERTIARY, "$A_{atm}$ — atmospheric"),
-        (result.a_gr, _C_QUATERNARY, "$A_{gr}$ — ground"),
-        (result.a_bar, _C_SECONDARY, "$A_{bar}$ — barrier"),
-    )
-    for values, color, label in terms:
-        term = np.asarray(values, dtype=np.float64)
-        bottom = np.where(term >= 0.0, pos_bottom, neg_bottom)
-        ax.bar(positions, term, bottom=bottom, color=color, label=label)
-        pos_bottom += np.maximum(term, 0.0)
-        neg_bottom += np.minimum(term, 0.0)
-
-    kwargs.setdefault("color", _C_REFERENCE)
-    kwargs.setdefault("marker", "D")
-    kwargs.setdefault("label", "$A$ — total")
-    ax.plot(positions, np.asarray(result.a_total, dtype=np.float64),
-            zorder=4, **kwargs)
-    ax.axhline(0.0, color=_C_MUTED, lw=0.8)
-    ax.set_ylabel("Attenuation A [dB]")
-    ax.set_title("ISO 9613-2 attenuation breakdown")
-    ax.legend(loc="best", fontsize="small")
-    ax.grid(True, axis="y", alpha=0.3)
-    return ax
 
 
 # ---------------------------------------------------------------------------
