+++ src/phonometry/_plotting.py
@@ -16,7 +16,7 @@ from ._plot.common import plot_aircraft_band_attenuation as plot_aircraft_band_a
 from ._plot.common import plot_ambient_noise as plot_ambient_noise
 from ._plot.building import plot_band_uncertainty as plot_band_uncertainty
 from ._plot.common import plot_bottom_loss as plot_bottom_loss
-from ._plot.common import plot_daily_exposure as plot_daily_exposure
+from ._plot.vibration import plot_daily_exposure as plot_daily_exposure
 from ._plot.room import plot_decay_curve as plot_decay_curve
 from ._plot.materials import plot_diffusion_polar as plot_diffusion_polar
 from ._plot.materials import plot_dynamic_stiffness as plot_dynamic_stiffness
@@ -38,16 +38,16 @@ from ._plot.building import plot_impact_insulation as plot_impact_insulation
 from ._plot.building import plot_impact_prediction as plot_impact_prediction
 from ._plot.building import plot_impact_rating as plot_impact_rating
 from ._plot.materials import plot_impedance_tube as plot_impedance_tube
-from ._plot.common import plot_impulse_prominence as plot_impulse_prominence
+from ._plot.environmental import plot_impulse_prominence as plot_impulse_prominence
 from ._plot.room import plot_impulse_response as plot_impulse_response
 from ._plot.materials import plot_insitu_absorption as plot_insitu_absorption
 from ._plot.building import plot_installed_structure_borne as plot_installed_structure_borne
 from ._plot.emission import plot_intensity as plot_intensity
-from ._plot.common import plot_mobility as plot_mobility
+from ._plot.vibration import plot_mobility as plot_mobility
 from ._plot.metrology import plot_monte_carlo as plot_monte_carlo
 from ._plot.psychoacoustics import plot_moore_glasberg_loudness as plot_moore_glasberg_loudness
 from ._plot.psychoacoustics import plot_moore_glasberg_time_loudness as plot_moore_glasberg_time_loudness
-from ._plot.common import plot_multiple_shock as plot_multiple_shock
+from ._plot.vibration import plot_multiple_shock as plot_multiple_shock
 from ._plot.hearing import plot_nipts as plot_nipts
 from ._plot.common import plot_noise_contour as plot_noise_contour
 from ._plot.room import plot_noise_criterion as plot_noise_criterion
@@ -55,7 +55,7 @@ from ._plot.common import plot_normal_modes as plot_normal_modes
 from ._plot.common import plot_npd_level as plot_npd_level
 from ._plot.hearing import plot_occupational_exposure as plot_occupational_exposure
 from ._plot.room import plot_open_plan as plot_open_plan
-from ._plot.common import plot_outdoor_attenuation as plot_outdoor_attenuation
+from ._plot.environmental import plot_outdoor_attenuation as plot_outdoor_attenuation
 from ._plot.common import plot_parabolic_equation as plot_parabolic_equation
 from ._plot.common import plot_pile_strike as plot_pile_strike
 from ._plot.psychoacoustics import plot_psychoacoustic_annoyance as plot_psychoacoustic_annoyance
@@ -75,16 +75,16 @@ from ._plot.common import plot_sound_speed_profile as plot_sound_speed_profile
 from ._plot.materials import plot_static_airflow as plot_static_airflow
 from ._plot.hearing import plot_sti as plot_sti
 from ._plot.building import plot_structure_borne_power as plot_structure_borne_power
-from ._plot.common import plot_tonal_adjustment as plot_tonal_adjustment
+from ._plot.environmental import plot_tonal_adjustment as plot_tonal_adjustment
 from ._plot.psychoacoustics import plot_tone_audibility as plot_tone_audibility
-from ._plot.common import plot_transfer_stiffness as plot_transfer_stiffness
+from ._plot.vibration import plot_transfer_stiffness as plot_transfer_stiffness
 from ._plot.common import plot_transmission_loss as plot_transmission_loss
 from ._plot.metrology import plot_uncertainty_budget as plot_uncertainty_budget
 from ._plot.building import plot_vibration_reduction as plot_vibration_reduction
 from ._plot.emission import plot_vibration_sound_power as plot_vibration_sound_power
-from ._plot.common import plot_vibration_weighting as plot_vibration_weighting
+from ._plot.vibration import plot_vibration_weighting as plot_vibration_weighting
 from ._plot.materials import plot_weighted_absorption as plot_weighted_absorption
 from ._plot.building import plot_weighted_rating as plot_weighted_rating
-from ._plot.common import plot_weighted_spectrum as plot_weighted_spectrum
-from ._plot.common import plot_wind_turbine_tonality as plot_wind_turbine_tonality
+from ._plot.vibration import plot_weighted_spectrum as plot_weighted_spectrum
+from ._plot.environmental import plot_wind_turbine_tonality as plot_wind_turbine_tonality
 from ._plot.psychoacoustics import plot_zwicker_loudness as plot_zwicker_loudness
