+++ src/phonometry/vibration/__init__.py
@@ -1,0 +1,140 @@
+#  Copyright (c) 2026. Jose M. Requena-Plens
+"""vibration domain of phonometry (see module docstrings)."""
+
+from __future__ import annotations
+
+from .human_vibration import (
+    DailyVibrationExposure,
+    ExposureAssessment,
+    HAV_EAV_A8,
+    HAV_ELV_A8,
+    HumanVibrationWarning,
+    REFERENCE_ACCELERATION,
+    REFERENCE_DURATION_S,
+    WBV_EAV_A8,
+    WBV_EAV_VDV,
+    WBV_ELV_A8,
+    WBV_ELV_VDV,
+    WEIGHTING_NAMES,
+    WeightedSpectrum,
+    WeightingResponse,
+    apply_weighting,
+    combine_partial_exposures,
+    crest_factor,
+    daily_exposure,
+    daily_vibration_exposure,
+    energy_equivalent_acceleration,
+    exposure_assessment,
+    frequency_weighting,
+    hav_daily_exposure,
+    hav_vwf_lifetime_years,
+    motion_sickness_dose_value,
+    mtvv,
+    partial_exposure,
+    running_rms,
+    vibration_dose_value,
+    vibration_total_value,
+    weighted_acceleration,
+    weighting_factors,
+)
+from .mechanical_mobility import (
+    FRF_UNITS,
+    MobilityResult,
+    convert_frf,
+    resonance_frequency,
+    sdof_accelerance,
+    sdof_mobility,
+    sdof_mobility_result,
+    sdof_receptance,
+)
+from .multiple_shock_vibration import (
+    MultipleShockResult,
+    acceleration_dose,
+    compression_dose,
+    daily_dose,
+    daily_dose_multi,
+    dose_from_peaks,
+    injury_probability,
+    injury_risk,
+    multiple_shock_assessment,
+    response_peaks,
+    seat_to_spine_transfer,
+    spinal_response,
+    static_stress,
+    ultimate_strength,
+)
+from .transfer_stiffness import (
+    REFERENCE_STIFFNESS,
+    TransferStiffnessResult,
+    base_transmissibility,
+    indirect_transfer_stiffness_result,
+    loss_factor,
+    transfer_stiffness_direct,
+    transfer_stiffness_indirect,
+    transfer_stiffness_level,
+)
+
+__all__ = [
+    "DailyVibrationExposure",
+    "ExposureAssessment",
+    "FRF_UNITS",
+    "HAV_EAV_A8",
+    "HAV_ELV_A8",
+    "HumanVibrationWarning",
+    "MobilityResult",
+    "MultipleShockResult",
+    "REFERENCE_ACCELERATION",
+    "REFERENCE_DURATION_S",
+    "REFERENCE_STIFFNESS",
+    "TransferStiffnessResult",
+    "WBV_EAV_A8",
+    "WBV_EAV_VDV",
+    "WBV_ELV_A8",
+    "WBV_ELV_VDV",
+    "WEIGHTING_NAMES",
+    "WeightedSpectrum",
+    "WeightingResponse",
+    "acceleration_dose",
+    "apply_weighting",
+    "base_transmissibility",
+    "combine_partial_exposures",
+    "compression_dose",
+    "convert_frf",
+    "crest_factor",
+    "daily_dose",
+    "daily_dose_multi",
+    "daily_exposure",
+    "daily_vibration_exposure",
+    "dose_from_peaks",
+    "energy_equivalent_acceleration",
+    "exposure_assessment",
+    "frequency_weighting",
+    "hav_daily_exposure",
+    "hav_vwf_lifetime_years",
+    "indirect_transfer_stiffness_result",
+    "injury_probability",
+    "injury_risk",
+    "loss_factor",
+    "motion_sickness_dose_value",
+    "mtvv",
+    "multiple_shock_assessment",
+    "partial_exposure",
+    "resonance_frequency",
+    "response_peaks",
+    "running_rms",
+    "sdof_accelerance",
+    "sdof_mobility",
+    "sdof_mobility_result",
+    "sdof_receptance",
+    "seat_to_spine_transfer",
+    "spinal_response",
+    "static_stress",
+    "transfer_stiffness_direct",
+    "transfer_stiffness_indirect",
+    "transfer_stiffness_level",
+    "ultimate_strength",
+    "vibration_dose_value",
+    "vibration_total_value",
+    "weighted_acceleration",
+    "weighting_factors",
+]
